VFB-aanpassingen naar eigen bestand
