+++ includes/system/siw-system.php
@@ -48,33 +48,6 @@ function siw_custom_upload_mimes( $existing_mimes=array() ){
 	return $existing_mimes;
 }
 
-//attachment verwijderen nadat deze per mail verstuurd is.
-add_action( 'vfbp_after_email', 'siw_delete_attachment_after_mail', 10, 2 );
-function siw_delete_attachment_after_mail( $entry_id, $form_id ) {
-	global $wpdb;
-
-	$attachments_args = array(
-		'post_type' 	=> 'attachment',
-		'post_parent'	=> $entry_id,
-		'fields' 		=> 'ids'
-	);
-	$attachments = get_posts( $attachments_args ); 
-	foreach ( $attachments as $attachment ) {
-		$attachment_url = wp_get_attachment_url( $attachment );
-		wp_delete_attachment( $attachment );
-		
-		$wpdb->query(
-			$wpdb->prepare(
-				"UPDATE $wpdb->postmeta
-				SET meta_value = 'gemaild'
-					WHERE post_id = %d
-					AND meta_value = %s;",
-				$entry_id,
-				$attachment_url
-			)
-		);
-	}
-}
 
 
 /*disable emojis*/
@@ -99,8 +72,6 @@ function siw_disable_emojicons_tinymce( $plugins ) {
 	}
 }
 
-
-
 /*instellen starttijd Updraft Plus backup*/
 add_filter('updraftplus_schedule_firsttime_db', 'siw_backup_time_db');
 add_filter('updraftplus_schedule_firsttime_files', 'siw_backup_time_files');
@@ -116,25 +87,6 @@ function siw_backup_time_files(){
 	$time = strtotime( 'tomorrow ' . $files_backup_time );
 	return $time;
 }
-
-
-/*
-VFB-pro aanpassingen
-*/
-
-add_action('wp_enqueue_scripts', 'siw_vfb_pro_scripts');
-function siw_vfb_pro_scripts(){
-	global $wp_scripts;
-	if ( $wp_scripts->registered['vfbp-js'] ){
-		$wp_scripts->registered['vfbp-js']->src = get_stylesheet_directory_uri() . '/assets/js/vfb-pro/vfb-js.min.js';
-	}
-	if ( $wp_scripts->registered['jquery-intl-tel'] ){
-		wp_register_script( 'jquery-phone-format', VFB_PLUGIN_URL . "public/assets/js/vendors/phone-format.min.js",array(), null, true);	
-		$wp_scripts->registered['jquery-intl-tel']->deps[] = 'jquery-phone-format';
-	}
-}
-
-
 
 /*dns-prefetch*/
 add_action('wp_head','siw_dns_prefetch', 0);
@@ -166,47 +118,7 @@ function siw_oembed_response_data( $data ){
 	return $data;
 }
 
-//cd-opties
 
-function siw_update_community_day_options(){
-
-	//haal cd-datums op
-	for ($x = 1 ; $x <= 9; $x++) {
-		$community_days[]= get_option("siw_community_day_{$x}");
-	}
-	asort( $community_days );
-	$hide_form_days_before_cd = siw_get_hide_form_days_before_cd();
-	$hide_option_after = date("Y-m-d", time() + ( $hide_form_days_before_cd * 24 * 60 * 60) );
-	
-	foreach($community_days as $community_day => $community_day_date) {
-		if( $community_day_date > $hide_option_after ){
-			$future_community_days[]['label']= siw_get_date_in_text( $community_day_date, false);
-		}
-	}
-
-	//zoek cd-formuliervraag
-	$field_id = siw_get_vfb_field_id('community_day_datums');
-	
-	global $wpdb;
-	if ( !isset($wpdb->vfbp_fields) ) {
-		$wpdb->vfbp_fields = $wpdb->prefix . 'vfbp_fields';
-	}
-	
-	$query = "SELECT $wpdb->vfbp_fields.data
-				FROM $wpdb->vfbp_fields
-				WHERE $wpdb->vfbp_fields.id = %d";
-	
-	$data = $wpdb->get_var( $wpdb->prepare( $query, $field_id));
-	$data = maybe_unserialize( $data );
-	
-	//update formuliervraag
-	$data['options'] = $future_community_days;
-	$query = "update $wpdb->vfbp_fields set $wpdb->vfbp_fields.data = %s where $wpdb->vfbp_fields.id = %d;";
-	$wpdb->query(
-		$wpdb->prepare( $query, maybe_serialize( $data ), $field_id )
-	);
-	
-}
 
 //hulpfunctie t.b.v. logging
 function siw_log( $content ){
